pyproject.toml
[project]
name = "mountain-recoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "build>=1.3.0",
    "click>=8.2.1",
    "django>=5.2.5",
    "django-bootstrap5>=25.2",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]

[project.scripts]
climb_rec = "app.cli:cli"
